# Refactor and optimize list db