--- a/src/core/api/api_db.pl
+++ b/src/core/api/api_db.pl
@@ -1,5 +1,6 @@
 :- module(api_db, [
-              list_databases/4,
+              list_databases/3,
+              list_existing_databases/2,
               pretty_print_databases/1
           ]).
 
@@ -8,6 +9,22 @@
 :- use_module(core(account)).
 :- use_module(core(query)).
 :- use_module(core(transaction)).
+
+get_all_databases(System_DB, Databases) :-
+    create_context(System_DB, Context),
+    findall(
+        Path,
+        (
+            ask(Context,
+                (
+                    t(Organization_Uri, system:resource_name, Organization^^xsd:string),
+                    t(Organization_Uri, rdf:type, system:'Organization'),
+                    t(Organization_Uri, system:resource_includes, Db_Uri),
+                    t(Db_Uri, rdf:type, system:'Database'),
+                    t(Db_Uri, system:resource_name, Name^^xsd:string)
+            )),
+            format(string(Path),"~s/~s",[Organization,Name])),
+        Databases).
 
 get_user_databases(System_DB, Auth, User_Databases) :-
     create_context(System_DB, Context),
@@ -22,19 +39,21 @@ get_user_databases(System_DB, Auth, User_Databases) :-
         (   member(DB,Scope),
             get_dict('@type',DB,'system:Database'),
             get_dict('@id',DB,ID),
+            Name = DB.'system:resource_name'.'@value',
             prefix_expand(ID, Prefixes, Ex_ID),
-            organization_database_name_uri(Context, Organization, Name, Ex_ID),
+            db_uri_organization(Context, Ex_ID, Organization),
             format(string(Path),"~s/~s",[Organization,Name])),
         User_Databases).
 
-list_databases(System_DB ,Auth, Databases, Database_Objects) :-
+list_databases(System_DB, Auth, Database_Objects) :-
+    (   is_super_user(Auth)
+    ->  get_all_databases(System_DB, User_Databases)
+    ;   get_user_databases(System_DB, Auth, User_Databases)),
+    list_existing_databases(User_Databases, Database_Objects).
 
-    (   Databases = []
-    ->  get_user_databases(System_DB, Auth, User_Databases)
-    ;   User_Databases = Databases),
-
+list_existing_databases(Databases, Database_Objects) :-
     findall(Database_Object,
-            (   member(DB, User_Databases),
+            (   member(DB, Databases),
                 list_database(DB, Database_Object)),
             Database_Objects).
 
@@ -86,9 +105,22 @@ test(list_all,
       cleanup(teardown_temp_store(State))]) :-
 
     super_user_authority(Auth),
-    list_databases(system_descriptor{}, Auth, [], Database_Objects),
+    list_databases(system_descriptor{}, Auth, Database_Objects),
     Expected_Objects = [_{branch_name:["main"],database_name:"admin/bar"},
                         _{branch_name:["main"],database_name:"admin/foo"}],
+
+    forall(member(Object, Database_Objects),
+           member(Object, Expected_Objects)).
+
+test(list_existing,
+     [setup((setup_temp_store(State),
+             create_db_without_schema("admin","foo2"),
+             create_db_without_schema("admin","bar2"))),
+      cleanup(teardown_temp_store(State))]) :-
+
+    list_existing_databases(["admin/foo2", "admin/bar2"], Database_Objects),
+    Expected_Objects = [_{branch_name:["main"],database_name:"admin/bar2"},
+                        _{branch_name:["main"],database_name:"admin/foo2"}],
 
     forall(member(Object, Database_Objects),
            member(Object, Expected_Objects)).